QATlib: 23.08.0 release
Changes from 23.02.0 to 23.08.0:
 * Removal of following insecure algorithms: Diffie-Hellman
   and Elliptic curves less than 256-bits.
 * Additional configuration profiles, including sym which
   facilitates improved symmetric crypto performance.
 * DC Chaining (Hash then compress).
 * Bug Fixes. See Resolved section in README.md

+++ quickassist/utilities/osal/src/linux/user_space/OsalServices.c
@@ -451,7 +451,7 @@ OSAL_PUBLIC OSAL_STATUS osalSleep(UINT32 milliseconds)
 {
     struct timespec resTime, remTime;
     INT32 status;
-    UINT16 mil_rem;
+    UINT32 mil_rem;
 
     /* Divide by number of millisec per second */
     resTime.tv_sec = milliseconds / OSAL_THOUSAND;
